+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # Changelog
 
 ## [Unreleased]
+### Changed
+- Client version updated on [5.0.16](https://github.com/reportportal/client-java/releases/tag/5.0.16)
+### Fixed
+- Empty interrupted suite in case of duplicate step
 
 ## [5.0.1]
 ### Changed